backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
google-auth==2.38.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.8.2
mypy_extensions==1.1.0
openai==1.64.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pipdeptree==2.26.1
platformdirs==4.3.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMuPDF==1.26.0
python-decouple==3.8
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
